third_party/WebKit/LayoutTests/http/tests/svg/as-object/intrinsic-size-change-embedded-object.html

<!doctype HTML>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<div style="width: 400px; height: 300px;">
  <object id=target style="border: 1px solid black"
      data="http://localhost:8000/svg/as-object/resources/objectcontents.svg"></object>
</div>
<script>
var t = async_test("objectsize");
onload = t.step_func_done(function() {
    let rect = target.getBoundingClientRect();
    assert_equals(rect.width, 400);
    assert_equals(rect.height, 798);
});
</script>
